Fix models

+++ AudioPlayerWebAPI/Apis/UserApi.cs
@@ -1,4 +1,6 @@
-﻿namespace AudioPlayerWebAPI.Apis
+﻿using AudioPlayerWebAPI.Models;
+
+namespace AudioPlayerWebAPI.Apis
 {
     public class UserApi : IApi
     {
@@ -39,14 +41,24 @@
         [Authorize]
         private async Task<IResult> Put([FromBody] User user)
         {
+            var validation = await _validator.ValidateAsync(user);
+            if (!validation.IsValid)
+            {
+                return Results.ValidationProblem(validation.ToDictionary());
+            }
             await _repository.UpdateUserAsync(user);
             await _repository.SaveAsync();
             return Results.Ok();
         }
 
         [Authorize]
-        private async Task<IResult> Delete(Guid userId)
+        private async Task<IResult> Delete(Guid userId, string password)
         {
+            var user = await _repository.GetUserByIdAsync(userId);
+            if (user.Password != Hash.GetSha1Hash(password))
+            {
+                return Results.Forbid();
+            }
             await _repository.DeleteUserAsync(userId);
             await _repository.SaveAsync();
             return Results.Ok();
